+++ packages/headless/src/controllers/search-parameter-manager/headless-search-parameter-manager.ts
@@ -1,6 +1,8 @@
 import {RecordValue, Schema} from '@coveo/bueno';
 import {Engine} from '../../app/headless-engine';
 import {getAdvancedSearchQueriesInitialState} from '../../features/advanced-search-queries/advanced-search-queries-state';
+import {partitionIntoParentsAndValues} from '../../features/facets/category-facet-set/category-facet-utils';
+import {FacetValueRequest} from '../../features/facets/facet-set/interfaces/request';
 import {getDebugInitialState} from '../../features/debug/debug-state';
 import {getPaginationInitialState} from '../../features/pagination/pagination-state';
 import {getQueryInitialState} from '../../features/query/query-state';
@@ -67,6 +69,8 @@ export function buildSearchParameterManager(
         ...getFirstResult(state),
         ...getNumberOfResults(state),
         ...getSortCriteria(state),
+        ...getFacets(state),
+        ...getCategoryFacets(state),
         ...getDebug(state),
       };
 
@@ -147,7 +151,47 @@ function getSortCriteria(state: Partial<SearchParametersState>) {
   return shouldInclude ? {sortCriteria} : {};
 }
 
+function getFacets(state: Partial<SearchParametersState>) {
+  if (state.facetSet === undefined) {
+    return {};
+  }
+
+  const f = Object.entries(state.facetSet)
+    .map(([facetId, request]) => {
+      const selectedValues = getSelectedValues(request.currentValues);
+      return selectedValues.length ? {[facetId]: selectedValues} : {};
+    })
+    .reduce((acc, obj) => ({...acc, ...obj}), {});
+
+  return Object.keys(f).length ? {f} : {};
+}
+
+function getSelectedValues(values: FacetValueRequest[]) {
+  return values.filter((fv) => fv.state === 'selected').map((fv) => fv.value);
+}
+
+function getCategoryFacets(state: Partial<SearchParametersState>) {
+  if (state.categoryFacetSet === undefined) {
+    return {};
+  }
+
+  const cf = Object.entries(state.categoryFacetSet)
+    .map(([facetId, request]) => {
+      const {parents} = partitionIntoParentsAndValues(request.currentValues);
+      const selectedValues = parents.map((p) => p.value);
+
+      return selectedValues.length ? {[facetId]: selectedValues} : {};
+    })
+    .reduce((acc, obj) => ({...acc, ...obj}), {});
+
+  return Object.keys(cf).length ? {cf} : {};
+}
+
 function getDebug(state: Partial<SearchParametersState>) {
+  if (state.debug === undefined) {
+    return {};
+  }
+
   const debug = state.debug;
   const shouldInclude = debug !== getDebugInitialState();
   return shouldInclude ? {debug} : {};