+++ packages/headless/src/test/mock-search-parameters.ts
@@ -11,6 +11,8 @@ export function buildMockSearchParameters(
     firstResult: 0,
     numberOfResults: 0,
     sortCriteria: '',
+    f: {},
+    cf: {},
     debug: false,
     ...config,
   };